ESLint - Extract Action to separate file to prevent dep cycle
