+++ src/utilities/logging/types.ts
@@ -1,0 +1,1 @@
+export type Action = (message: string | Error, details?: Record<string, unknown>) => void;
